requirements.txt
# requirements.txt
flask
google-generativeai
python-dotenv
